Include zip_slip tests into the main test file
